+++ lib/config.rb
