--- lib/soundbot/scripts/record.sh
@@ -1,10 +1,0 @@
-#!/usr/bin/env bash
-
-
-
-# manage zita-a2j process...
-
-#TODO: set recording directory in config
-
-timemachine -n 'powermic' -s -f wav -a -b -10.0 -e -20.0 -T 5 -t 10 \
--p "/home/b08x/Studio/timemachine/something-" powermic:capture_1
