+++ lib/util.rb
